mudança do voto para double e retornando todos os tipos de votos em alldata

+++ FeiraPreta/Domain/EventScore.cs
@@ -9,7 +9,7 @@ namespace FeiraPreta.Domain
     {
         public Guid Id { get; set; }
 
-        public float Value { get; set; }
+        public double Value { get; set; }
 
         public DateTime CreatedDate { get; set; }
     }
